Refactor FeedbackPanelMentorDiscussion
