+++ app/javascript/components/editor/FeedbackPanel/FeedbackPanelMentoringDiscussion/index.ts
@@ -1,0 +1,4 @@
+export { MentoringDiscussion } from './FeedbackPanelMentoringDiscussion'
+export { ReadonlyDiscussionPostView } from './ReadonlyDiscussionPostView'
+export { ReadonlyIterationMarker } from './ReadOnlyIterationMarker'
+export { PendingMentoringRequest } from './PendingMentoringRequest'
